requirements.txt
gradio>=3.0
google-genai>=0.10.0
python-dotenv
Pillow
